E-BOOK ANALYSIS AND REPRESENTATION
Program downloads e-books from (Wikibooks or Wikisource) and save them to a text file. The name of the book will be taken as user input. After correctly saving the e-book to a text file, program be able to read it and create word frequencies of that book, meaning counting the number of times words has been used in that e-book. During this operation, program also remove stop words (the, a, he, she, it, etc.), words that are meaningless by themselves but usually have the highest frequency of available words.